refactor: organise radio namespace by protocol family
