+++ opendis/record/symbolic_names.py
@@ -1,0 +1,136 @@
+"""6.1.8 Table 25: Constants for DIS7 symbolic names"""
+
+AGG_RESPONSE_DFLT = 10  # Default: 10 s
+ALL_AGGS = 0xFFFF  # Hexadecimal
+ALL_APPLIC = 0xFFFF  # Hexadecimal
+ALL_BEAMS = 0xFF  # Hexadecimal
+ALL_EMITTERS = 0xFF  # Hexadecimal
+ALL_ENTITIES = 0xFFFF  # Hexadecimal
+ALL_OBJECTS = 0xFFFF  # Hexadecimal
+ALL_SITES = 0xFFFF  # Hexadecimal
+COLLISION_ELASTIC_TIMEOUT = 5  # Default: 5 s
+COLLISION_THRSH = 0.1  # Default: 0.1 m/s
+DE_AREA_AIMING_THRSH = 10  # Default: 10°
+DE_ENERGY_THRSH = 1.0  # Default: 1.0%
+DE_PRECISION_AIMING_THRSH = 0.5  # Default: 0.5 m
+DRA_ORIENT_THRSH = 3  # Default: 3°
+DRA_POS_THRSH = 1  # Default: 1 m
+D_SPOT_NO_ENTITY = None  # No entity
+EE_AD_PULRAT_THRSH = 0.017  # Default: 0.017 rad/s
+EE_AD_PULACC_THRSH = 0.017  # Default: 0.017 rad/s²
+EE_AZ_THRSH = 1  # Default: 1°
+EE_EL_THRSH = 1  # Default: 1°
+EE_ERP_THRSH = 1.0  # Default: 1.0 dBm
+EE_FREQ_THRSH = 1  # Default: 1 Hz
+EE_FRNG_THRSH = 1  # Default: 1 Hz
+EE_FT_VEL_THRSH = 1.0  # Default: 1.0 m/s
+EE_FT_ACC_THRSH = 1.0  # Default: 1.0 m/s²
+EE_FT_MWD_THRSH = 10000  # Default: 10000 m
+EE_FT_KT_THRSH = 10  # Default: 10 s
+EE_FT_ESP_THRSH = 10  # Default: 10 m
+EE_HIGH_DENSITY_THRSH = 10  # Default: 10 entities/beam
+EE_PRF_THRSH = 1  # Default: 1 Hz
+EE_PW_THRSH = 1  # Default: 1
+ENTITY_ID_UNKNOWN = None  # No entity
+EP_DIMENSION_THRSH = 1  # Default: 1 m
+EP_NO_SEQUENCE = 0xFFFF  # Hexadecimal
+EP_POS_THRSH = 1  # Default: 1 m shift
+EP_STATE_THRSH = 10  # Default: ±10%
+GD_GEOMETRY_CHANGE = 10  # Default: ±10%
+GD_STATE_CHANGE = 10  # Default: ±10%
+HBT_DAMAGE_TIMEOUT_MPLIER = 2.4  # Default: 2.4
+HBT_ESPDU_KIND_CULTURAL_FEATURE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_ENVIRONMENTAL = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_EXPENDABLE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_LIFE_FORM = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_MUNITION = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_RADIO = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_SENSOR_EMITTER = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_KIND_SUPPLY = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_PLATFORM_AIR = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_PLATFORM_LAND = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_PLATFORM_SPACE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_PLATFORM_SUBSURFACE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_ESPDU_PLATFORM_SURFACE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_AGGREGATE_STATE = 30  # Default: 30 s, Tolerance: ±10%
+HBT_PDU_APPEARANCE = 60  # Default: 60 s, Tolerance: ±10%
+HBT_PDU_DE_FIRE = 0.5  # Default: 0.5 s, Tolerance: ±10%
+HBT_PDU_DESIGNATOR = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_EE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_ENTITY_DAMAGE = 10  # Default: 10 s, Tolerance: ±10%
+HBT_PDU_ENVIRONMENTAL_PROCESS = 15  # Default: 15 s, Tolerance: ±10%
+HBT_PDU_GRIDDED_DATA = 900  # Default: 15 min, Tolerance: ±10%
+HBT_PDU_IFF = 10  # Default: 10 s, Tolerance: ±10%
+HBT_PDU_ISGROUPOF = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_MINEFIELD_DATA = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_MINEFIELD_STATE = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_RECEIVER = 5  # Default: 5 s, Tolerance: ±10%
+HBT_PDU_SEES = 180  # Default: 3 min, Tolerance: ±10%
+HBT_PDU_TRANSMITTER = 2  # Default: 2 s, Tolerance: ±10%
+HBT_PDU_TSPI = 30  # Default: 30 s, Tolerance: ±10%
+HBT_PDU_UA = 180  # Default: 3 min, Tolerance: ±10%
+HBT_STATIONARY = 60  # Default: 1 min, Tolerance: ±10%
+HBT_TIMEOUT_MPLIER = 2.4  # Default: 2.4
+HQ_TOD_DIFF_THRSH = 20  # Default: 20 ms
+IFF_AZ_THRSH = 3  # Default: 3°
+IFF_CHG_LATENCY = 2  # Default: 2 s
+IFF_EL_THRSH = 3  # Default: 3°
+IFF_IP_REPLY_TIMER = 30  # Default: 30 s
+IFF_PDU_FINAL = 10  # Default: 10 s
+IFF_PDU_RESUME = 10  # Default: 10 s
+IO_UNTIL_FURTHER_NOTICE = 65535  # Fixed
+MAX_PDU_SIZE_BITS = 65536  # Fixed
+MAX_PDU_SIZE_OCTETS = 8192  # Fixed
+MINEFIELD_CHANGE = 2.5  # Default: 2.5 s
+MINEFIELD_RESPONSE_TIMER = 1  # Default: 1 s
+MULTIPLES_PRESENT = 0  # Fixed
+NO_AGG = 0  # Fixed
+NO_APPLIC = 0  # Fixed
+NO_BEAM = 0  # Fixed
+NO_CATEGORY = 0  # Fixed
+NO_EMITTER = 0  # Fixed
+NO_ENTITY = 0  # Fixed
+NO_FIRE_MISSION = 0  # Fixed
+NO_KIND = 0  # Fixed
+NO_OBJECT = 0  # Fixed
+NO_PATTERN = 0.0  # Fixed
+NO_REF_NUMBER = 0  # Fixed
+NO_SITE = 0  # Fixed
+NO_SPECIFIC = 0  # Fixed
+NO_SPECIFIC_ENTITY = None  # No entity
+NO_SUBCAT = 0  # Fixed
+NO_VALUE = 0  # Fixed
+NON_SYNC_THRSH = 60  # Default: 1 min
+POWER_ENGINE_OFF = -100.0  # Fixed
+POWER_IDLE = 0.0  # Fixed
+POWER_MAX_AFTERBURNER = 100.0  # Fixed
+POWER_MILITARY = 50.0  # Fixed
+POWER_MIN_AFTERBURNER = 51.0  # Fixed
+REPAR_REC_T1 = 5  # Default: 5 s
+REPAR_SUP_T1 = 12  # Default: 12 s
+REPAR_SUP_T2 = 12  # Default: 12 s
+RESUP_REC_T1 = 5  # Default: 5 s
+RESUP_REC_T2 = 55  # Default: 55 s
+RESUP_SUP_T1 = 60  # Default: 1 min
+RQST_ASSIGN_ID = 0xFFFE  # Hexadecimal
+SEES_NDA_THRSH = 2  # Default: ±2° in the axis of deflection
+SEES_PS_THRSH = 10  # Default: ±10% of the maximum value of the Power Setting
+SEES_RPM_THRSH = 5  # Default: ±5% of the maximum engine speed in RPM
+SMALLEST_MTU_OCTETS = 1400  # Default: 1400 octets for IPv4 networks
+SM_REL_RETRY_CNT = 3  # Default: 3
+SM_REL_RETRY_DELAY = 2  # Default: 2 s
+TARGET_ID_UNKNOWN = None  # No entity
+TIMESTAMP_AHEAD = 5  # Default: 5 s
+TIMESTAMP_BEHIND = 5  # Default: 5 s
+TI_TIMER1_DFLT = 2  # Default: 2 s
+TI_TIMER2_DFLT = 12  # Default: 12 s
+TO_AUTO_RESPONSE_TIMER = 5  # Default: 5 s
+TO_MAN_RESPONSE_TIMER = 120  # Default: 120 s
+TR_TIMER1_DFLT = 5  # Default: 5 s
+TR_TIMER2_DFLT = 60  # Default: 60 s
+TRANS_ORIENT_THRSH = 180  # Default: 180°
+TRANS_POS_THRSH = 500  # Default: 500 m
+UA_ORIENT_THRSH = 2  # Default: 2°
+UA_POS_THRSH = 10  # Default: 10 m
+UA_SRPM_ROC_THRSH = 10  # Default: ±10% of maximum rate of change
+UA_SRPM_THRSH = 5  # Default: ±5% of maximum shaft rate in RPM
